+++ std/ft_putnbr.c
@@ -3,16 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gsteyn <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: gsteyn <gsteyn@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/23 05:35:11 by gsteyn            #+#    #+#             */
-/*   Updated: 2018/05/29 11:30:31 by gsteyn           ###   ########.fr       */
+/*   Updated: 2018/09/04 12:19:26 by gsteyn           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <inttypes.h>
 
-void	ft_putnbr(int n)
+void	ft_putnbr(intmax_t n)
 {
 	if (n == -2147483648)
 		ft_putstr("-2147483648");
