+++ std/ft_putnbr_base.c
@@ -6,13 +6,13 @@
 /*   By: gsteyn <gsteyn@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/03 10:33:42 by gsteyn            #+#    #+#             */
-/*   Updated: 2018/08/06 13:16:17 by gsteyn           ###   ########.fr       */
+/*   Updated: 2018/09/04 12:22:43 by gsteyn           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void			ft_putnbr_base(long long n, int base)
+void			ft_putnbr_base(intmax_t n, int base)
 {
 	if (base < 2 || base > 16)
 		return ;
